Update index.html
Fixing a few things:
First, the clicking points do not map directly to the dogs in the image. Sometimes, I need to click around to get the correct location, which is sometimes not even where the dog is located.
Second, the areas where you can click are all gray circles, which are visible to the game user. Please do not make the areas where one can click visible.
Third, it is still possible to rack up multiple points for a single dog/region. Please correct this so that a user can only score points once for each correct answer